+++ app/src/main/java/org/simple/clinic/sync/SyncModule.kt
@@ -48,6 +48,8 @@ import org.simple.clinic.questionnaireresponse.di.QuestionnaireResponseModule
 import org.simple.clinic.questionnaireresponse.sync.QuestionnaireResponseSync
 import org.simple.clinic.reports.ReportsModule
 import org.simple.clinic.reports.ReportsSync
+import org.simple.clinic.returnscore.di.ReturnScoreModule
+import org.simple.clinic.returnscore.sync.ReturnScoreSync
 import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationSync
 import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecordRepository
 import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecordSync
@@ -74,7 +76,8 @@ import javax.inject.Named
   QuestionnaireModule::class,
   QuestionnaireResponseModule::class,
   PatientAttributeModule::class,
-  CVDRiskModule::class
+  CVDRiskModule::class,
+  ReturnScoreModule::class,
 ])
 class SyncModule {
 
@@ -99,6 +102,7 @@ class SyncModule {
       questionnaireResponseSync: QuestionnaireResponseSync,
       patientAttributeSync: PatientAttributeSync,
       cvdRiskSync: CVDRiskSync,
+      returnScoreSync: ReturnScoreSync
   ): List<ModelSync> {
     val optionalSyncs = if (features.isEnabled(Feature.CallResultSyncEnabled)) listOf(callResultSync) else emptyList()
 
@@ -106,7 +110,7 @@ class SyncModule {
         questionnaireSync, questionnaireResponseSync, protocolSync, reportsSync, helpSync,
         patientSync, bloodPressureSync, medicalHistorySync, appointmentSync, prescriptionSync,
         bloodSugarSync, teleconsultationMedicalOfficersSync,
-        teleconsultRecordSync, drugSync, patientAttributeSync, cvdRiskSync
+        teleconsultRecordSync, drugSync, patientAttributeSync, cvdRiskSync, returnScoreSync
     ) + optionalSyncs
   }